Prepare tensorflow_recommender to make ResourceVariable a CompositeTensor.
We are going to make ResourceVariable to be a CompositeTensor. As a result, tf.nest.flatten with expand_composites=True will expand ResourceVariables to be dt_resource tensors. So let's rely on layer.__setattr__, which calls layer._track_variables and is capable to handle Composite Variables, like ShardedVariable.

This CL shouldn't change the model behavior.

PiperOrigin-RevId: 467699678